rename model_pt to model
