+++ sockeye/checkpoint_decoder.py
@@ -28,7 +28,7 @@ from . import constants as C
 from . import data_io
 from . import evaluate
 from . import inference
-from . import model_pt
+from . import model
 from . import utils
 from . import vocab
