+++ sockeye/score.py
@@ -26,7 +26,7 @@ from . import data_io
 from . import utils
 from .beam_search import CandidateScorer
 from .log import setup_main_logger
-from .model_pt import load_model
+from .model import load_model
 from .output_handler import get_output_handler
 from .scoring import BatchScorer, Scorer
 from .utils import check_condition
