+++ test/integration/test_seq_copy_int.py
@@ -34,7 +34,7 @@ import sockeye.checkpoint_decoder
 import sockeye.evaluate
 from sockeye import constants as C
 from sockeye.config import Config
-from sockeye.model_pt import load_model
+from sockeye.model import load_model
 from sockeye.test_utils import run_train_translate, tmp_digits_dataset
 from test.common import check_train_translate
