+++ test/unit/test_model.py
@@ -18,7 +18,7 @@ import torch as pt
 
 import sockeye.constants as C
 import sockeye.data_io
-import sockeye.model_pt
+import sockeye.model
 import sockeye.transformer
